Change AddEvent component to dumb
